REF: fix ioc and docs
