+++ docs/Makefile
@@ -14,6 +14,10 @@ help:
 
 .PHONY: help Makefile
 
+clean:
+	@echo "Removing everything under 'build' and 'source/ioc'..."
+	@rm -rf build/ source/ioc
+
 # Catch-all target: route all unknown targets to Sphinx using the new
 # "make mode" option.  $(O) is meant as a shortcut for $(SPHINXOPTS).
 %: Makefile
