--- solid_attenuator/ioc_lfe_at2l0_calc/db/__init__.py
@@ -1,3 +1,0 @@
-from .system import SystemGroup
-
-__all__ = ['SystemGroup']
